MGDOBR-270: move CloudEvent parsing to ExecutorService
