--- executor/src/main/java/com/redhat/service/smartevents/executor/ExecutorsProvider.java
@@ -1,5 +1,0 @@
-package com.redhat.service.smartevents.executor;
-
-public interface ExecutorsProvider {
-    Executor getExecutor();
-}
