+++ executor/src/test/java/com/redhat/service/smartevents/executor/ExecutorServiceTest.java
@@ -1,66 +1,66 @@
 package com.redhat.service.smartevents.executor;
 
-import java.net.URI;
-
-import javax.inject.Inject;
+import java.util.stream.Stream;
 
 import org.eclipse.microprofile.reactive.messaging.Message;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.verification.VerificationMode;
 
-import com.redhat.service.smartevents.infra.models.dto.BridgeDTO;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.redhat.service.smartevents.infra.models.dto.ProcessorDTO;
-import com.redhat.service.smartevents.infra.utils.CloudEventUtils;
 
 import io.cloudevents.CloudEvent;
-import io.cloudevents.core.builder.CloudEventBuilder;
-import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
 
+import static com.redhat.service.smartevents.executor.ExecutorTestUtils.PLAIN_EVENT_JSON;
+import static com.redhat.service.smartevents.executor.ExecutorTestUtils.createCloudEventString;
+import static com.redhat.service.smartevents.executor.ExecutorTestUtils.createSinkProcessorWithResolvedAction;
+import static com.redhat.service.smartevents.executor.ExecutorTestUtils.createSinkProcessorWithSameAction;
+import static com.redhat.service.smartevents.executor.ExecutorTestUtils.createSourceProcessor;
+import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@QuarkusTest
-public class ExecutorServiceTest {
+class ExecutorServiceTest {
 
-    private static final String BRIDGE_ID = "bridgeId";
+    private static final String BROKEN_JSON = "/%=({}[][]";
 
-    @Inject
-    ExecutorsService executorsService;
+    @ParameterizedTest
+    @MethodSource("executorServiceTestArgs")
+    @SuppressWarnings("unchecked")
+    void test(ProcessorDTO processor, String inputEvent, VerificationMode wantedNumberOfOnEventInvocations) {
+        Executor executorMock = mock(Executor.class);
+        when(executorMock.getProcessor()).thenReturn(processor);
 
-    @InjectMock
-    ExecutorsProvider executorsProvider;
+        ExecutorService executorService = new ExecutorService();
+        executorService.executor = executorMock;
+        executorService.mapper = new ObjectMapper();
+        executorService.init();
 
-    Executor executor;
+        Message<String> inputMessage = mock(Message.class);
+        when(inputMessage.getPayload()).thenReturn(inputEvent);
 
-    @BeforeEach
-    public void before() {
-        executor = mock(Executor.class);
-
-        BridgeDTO bridgeDTO = mock(BridgeDTO.class);
-        when(bridgeDTO.getId()).thenReturn(BRIDGE_ID);
-
-        ProcessorDTO processorDTO = mock(ProcessorDTO.class);
-        when(processorDTO.getBridgeId()).thenReturn(BRIDGE_ID);
-
-        when(executor.getProcessor()).thenReturn(processorDTO);
-        when(executorsProvider.getExecutor()).thenReturn(executor);
+        assertThatNoException().isThrownBy(() -> executorService.processEvent(inputMessage));
+        verify(executorMock, wantedNumberOfOnEventInvocations).onEvent(any(CloudEvent.class));
+        verify(inputMessage).ack();
     }
 
-    @Test
-    public void handleEvent() {
-        CloudEvent cloudEvent = CloudEventBuilder
-                .v1()
-                .withId("foo")
-                .withSource(URI.create("bar"))
-                .withType("myType")
-                .build();
-
-        executorsService.processBridgeEvent(Message.of(CloudEventUtils.encode(cloudEvent)));
-
-        verify(executor, times(1)).onEvent(any(String.class));
+    private static Stream<Arguments> executorServiceTestArgs() {
+        Object[][] arguments = {
+                { createSourceProcessor(), BROKEN_JSON, never() },
+                { createSourceProcessor(), PLAIN_EVENT_JSON, times(1) },
+                { createSinkProcessorWithSameAction(), BROKEN_JSON, never() },
+                { createSinkProcessorWithSameAction(), PLAIN_EVENT_JSON, never() },
+                { createSinkProcessorWithSameAction(), createCloudEventString(), times(1) },
+                { createSinkProcessorWithResolvedAction(), BROKEN_JSON, never() },
+                { createSinkProcessorWithResolvedAction(), PLAIN_EVENT_JSON, never() },
+                { createSinkProcessorWithResolvedAction(), createCloudEventString(), times(1) }
+        };
+        return Stream.of(arguments).map(Arguments::of);
     }
 }
